Remove the org option
The only inputs we really need (so far) are token and remote. The rest
(like the org) we will get from git commands. Because we will be listing
branches with `git branch -r`, we need to be in a git directory anyway.

Specifically, we will get the org from the remote URL, but we will need
the remote name to figure out which remote to get the URL of.