Prevent squashed-diff sh from failing
If the last thing that the Bash commands underlying --squashed-diff
looked for in its while loop failed the [[ ]] test, the command would
exit non-zero, causing an error in the Node process. This change adds
`true` to the end of each loop iteration, so the command does not exit
non-zero in this non-error condition.